requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
click==8.1.7
dictdiffer==0.9.0
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
packaging==23.2
pluggy==1.3.0
pydantic==2.5.1
pydantic_core==2.14.3
pypika-tortoise==0.1.6
pytest==7.4.3
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.3
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
